backend/requirements.txt
# Web Framework & Server
fastapi==0.115.0
uvicorn[standard]==0.30.0
starlette==0.38.1

# Pydantic & Validation
pydantic[email]==2.12.4
pydantic-settings==2.4.0
pydantic-core==2.41.5
email-validator==2.3.0

# Type Hints
typing-extensions==4.14.1

# HTTP Client Stack
httpx==0.27.0
httpcore==1.0.9
anyio==4.4.0
sniffio==1.3.1
certifi==2024.12.14
idna==3.7
charset-normalizer==3.4.4

# Auth & Security
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
tenacity==9.1.2
jinja2==3.1.4
python-dotenv==1.0.1
bcrypt==4.0.1

# AWS / R2
boto3==1.34.141
botocore==1.34.141
jmespath==1.0.1
s3transfer==0.10.2
urllib3==2.2.2

# Database / ORM
psycopg==3.2.1
prisma==0.12.0

# LLM APIs
google-generativeai==0.8.3
groq==0.11.0

# Multimedia & Presentation
python-pptx==0.6.21
pillow==11.0.0
moviepy==1.0.3
imageio==2.37.0
imageio-ffmpeg==0.6.0
pydub==0.25.1
elevenlabs==2.22.0
reportlab==4.2.0

# Utilities
numpy==2.1.3
requests==2.32.3
tqdm==4.67.0
decorator==4.4.2
click==8.1.7
colorama==0.4.6
proglog==0.1.12
websockets==14.1
reportlab==4.2.2         # for enhanced PDF generation
networkx==3.4.2          # for flowcharts and mindmaps
matplotlib==3.9.2        # visualization backend
graphviz==0.20.3         # mindmap and diagram export support
colorama==0.4.6          # terminal color output for builder scripts
supabase==2.6.0
